% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{FastLaplace}
\alias{FastLaplace}
\title{Implements the fast Laplace algorithm.}
\usage{
FastLaplace(PHI, y, sigma2, eta, verbose)
}
\arguments{
\item{PHI}{measurement matrix.}

\item{y}{CS measurements.}

\item{sigma2}{Initial noise variance.}

\item{eta}{Threshold in determining convergence of marginal likelihood.}

\item{verbose}{Print to screen which basis are added, re-estimated, or deleted.}
}
\value{
A list containing the following elements:
\tabular{lll}{
  \code{weights} \tab \tab sparse weights\cr
  \code{used} \tab \tab the positions of sparse weights\cr
  \code{sigma2} \tab \tab re-estimated noise variance\cr
  \code{errbars} \tab \tab one standard deviation around the sparse weights\cr
  \code{alpha} \tab \tab sparse hyperparameters (1/gamma)
}
}
\description{
This code implements the fast Laplace algorithm from [1], which is based on
  the BCS code available from [2].
}
\references{
[1] S. D. Babacan, R. Molina, A. K. Katsaggelos. "Bayesian Compressive
Sensing using Laplace Priors",

[2] S. Ji, Y. Xue, L. Carin, "Bayesian Compressive Sensing," IEEE Trans.
Signal Processing, vol. 56, no. 6, June 2008.

[3] M. Tipping and A. Faul, “Fast marginal likelihood maximisation
for sparse Bayesian models,” in Proc. 9th Int. Workshop Artificial Intelligence
and Statistics, C. M. Bishop and B. J. Frey, Eds., 2003.
}

