% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{FourierBasis}
\alias{FourierBasis}
\title{Finds the discrete Fourier transformation matrix.}
\usage{
FourierBasis(train, tlist, M, N)
}
\arguments{
\item{PHI}{Measurement matrix.}

\item{y}{Sample from original signal.}

\item{eta}{Tolerance level in determining convergence of marginal likelihood.}

\item{verbose}{Print to screen which basis are added, re-estimated, or deleted.}
}
\value{
The sparse signal as found by the fast Laplace algorithm.
}
\description{
Given samples from some signal and given a measurement matrix, find the
sparse respresentation of a signal using compressive sensing.
}

