% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{WaveletBasis}
\alias{WaveletBasis}
\title{Finds the discrete wavelet transform matrix.}
\usage{
WaveletBasis(signal, N, train, wavelet)
}
\arguments{
\item{signal}{original signal to find the wavelet transform for.}

\item{N}{number of wavelet basis to keep.}

\item{train}{the rows to keep of the matrix.}

\item{wavelet}{the type of wavelet basis to use. See the documentation for the
package \code{wmtsa} for more information.}
}
\value{
A PxN discrete wavelet transform matrix, where P is equal to the
       length of \code{train} and N is the number of basis.
}
\description{
Uses the package \code{wmtsa} to find the matrix of a wavelet basis. Each
column is a basis function from the given basis, evaluated at particular points
along the rows of the matrix.
}

