% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{BSplineBasis}
\alias{BSplineBasis}
\title{Finds the transformation matrix for the B-spline basis.}
\usage{
BSplineBasis(tlist, N, train = NULL)
}
\arguments{
\item{tlist}{an array of the specific points where the basis are evaluated at.}

\item{N}{number of basis in the matrix.}

\item{train}{indices corresponding to which rows of the matrix to keep.
Default is to keep all rows.}
}
\value{
A PxN matrix where P is equal to the length of \code{train} and N is
       the number of B-spline basis.
}
\description{
Uses the package \code{\link{fda}} to find a transformation matrix where each
column is a basis function from the B-spline basis, evaluated at specified
points along the rows of the matrix.
}

