% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{FourierBasis}
\alias{FourierBasis}
\title{Finds the discrete Fourier transformation matrix.}
\usage{
FourierBasis(tlist, N, train = NULL)
}
\arguments{
\item{tlist}{an array of the specific points where the basis are evaluated at.}

\item{N}{number of basis in the matrix.}

\item{train}{indices corresponding to which rows of the matrix to keep.}
}
\value{
A PxN discrete Fourier transformation matrix where P is equal to the
       length of \code{train} and N is the number of basis.
}
\description{
Uses the package \code{fda} to find a matrix where the columns are the different
basis and the rows are where the basis are evaluated at.
}

