% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{WaveletBasis}
\alias{WaveletBasis}
\title{Finds the discrete wavelet transform matrix.}
\usage{
WaveletBasis(train, N, signal, wavelet)
}
\arguments{
\item{PHI}{Measurement matrix.}

\item{y}{Sample from original signal.}

\item{eta}{Tolerance level in determining convergence of marginal likelihood.}

\item{verbose}{Print to screen which basis are added, re-estimated, or deleted.}
}
\value{
The sparse signal as found by the fast Laplace algorithm.
}
\description{
Finds the discrete wavelet transform matrix.
}

