% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.R
\name{WaveletBasis}
\alias{WaveletBasis}
\title{Finds the discrete wavelet transform matrix.}
\usage{
WaveletBasis(N, train = NULL, wavelet = "Haar")
}
\arguments{
\item{N}{number of wavelet basis to keep.}

\item{train}{indices corresponding to which rows of the matrix to keep.
Default is to keep all rows.}

\item{wavelet}{the type of wavelet basis to use. See
\code{\link[wmtsa]{wavDaubechies}} from \code{wmtsa} for types.}
}
\value{
A PxN discrete wavelet transform matrix, where P is equal to the
       length of \code{train} and N is the number of basis.
}
\description{
Uses the functions \code{\link[wmtsa]{wavDWT}} and
\code{\link[wmtsa]{reconstruct}} from the \code{wmtsa} package to find the
transformation matrix of the given wavelet basis type. Each column of the
matrix is a basis function from the wavelet basis type, evaluated at specified
points along the rows of the matrix.
}

